all: refactor select lb policy from a list of raw configs
